refactor: Keep GQL-related tests in a folder
