+++ k6/index.ts
@@ -1,1 +1,4 @@
-export * from "./components";
+// @ts-nocheck - TS complains about duplicate options exports. This is necessary
+// to have one entry point to generate all JS files.
+export * from "./graphql/components";
+export * from "./graphql/observations";
